fix: avoid checking prividium env at runtime
