+++ packages/api/src/config/index.ts
@@ -2,6 +2,7 @@ import { TypeOrmModuleOptions } from "@nestjs/typeorm";
 import * as featureFlags from "./featureFlags";
 import { BASE_TOKEN_L1_ADDRESS, BASE_TOKEN_L2_ADDRESS } from "../common/constants";
 import { z } from "zod";
+import { prividium } from "./featureFlags";
 
 export type BaseToken = {
   name: string;
@@ -149,23 +150,10 @@ export default () => {
   };
 
   const getPrividiumConfig = () => {
-    if (!featureFlags.prividium) {
-      return null;
-    }
-
-    const schema = z.object({
-      privateRpcUrl: z.string().url(),
-      privateRpcSecret: z.string().min(1),
-    });
-    const result = schema.safeParse({
+    return {
       privateRpcUrl: PRIVIDIUM_PRIVATE_RPC_URL,
       privateRpcSecret: PRIVIDIUM_PRIVATE_RPC_SECRET,
-    });
-    if (!result.success) {
-      throw new Error("Invalid Prividium config");
-    }
-
-    return result.data;
+    };
   };
 
   return {
